Lua NPC Scheduling
- Rewrite the C++ NPC scheduler in Lua, to allow and simplify future changes.
- Add a Lua-based NPC configuration file to replace the current XML format.
- Remove Simulatable_Command & allow every Command to have is_complete(ticks).
- Expose is_complete(ticks) and execute() operations for Lua's Command_Result.
- Rename Map::add_object creation overload to add_new_object for clarity.
- Expose add_object and add_new_object Map functions in Lua.
- Allow setting object state in Lua.
- Add Wait_Command and Move_To_Command creation functions to Lua.
- When moving to a keypoint, set object direction to keypoint's when reached.
- Remove the C++ NPC classes and references to them.
- Add npcs as a global Lua table with all NPCs keyed by name.
- Add a Lua function get_map_npcs to replace the old game:get_map_npcs.

+++ src/commands/wait_command.cpp
@@ -1,7 +1,7 @@
 #include "../../include/commands/wait_command.hpp"
 #include "../../include/game.hpp"
 
-Wait_Command::Wait_Command(Game& game, long duration, int start)
+Wait_Command::Wait_Command(Game& game, int duration, int start)
 	: game(game), duration(duration),
 	start_time(start < 0 ? game.ticks() : start) {}
